refactor(doctor): extract check modules
